Added script to setup, test in devcloud.
